Issue #6728 - QUIC and HTTP/3
- Implemented stream idle timeouts.

+++ jetty-quic/quic-common/src/main/java/org/eclipse/jetty/quic/common/ProtocolSession.java
@@ -120,7 +120,7 @@ public abstract class ProtocolSession
         connection.onOpen();
     }
 
-    public boolean close(int error, String reason)
+    public boolean close(long error, String reason)
     {
         return getQuicSession().close(error, reason);
     }
